% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehsr-package.R
\docType{package}
\name{ehsr-package}
\alias{ehsr}
\alias{ehsr-package}
\title{ehsr: Imports Saved Data from the English Housing Survey}
\description{
Functions to easily import microdata from English Housing Survey datasets that have been downloaded from the UK Data Service. When provided with a folder where the EHS datasets are saved (including in sub-folders), ehsr will identify and import data for the years you specify, distinguishing between the household and housing stock datasets. It will also standardise variable names and value labels where these have evolved over time.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jgleeson/ehsr}
  \item Report bugs at \url{https://github.com/jgleeson/ehsr/issues}
}

}
\author{
\strong{Maintainer}: James Gleeson \email{jimgleeson@gmail.com}

}
\keyword{internal}
