% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehs_sl.R
\name{hh_adapt}
\alias{hh_adapt}
\title{Import data on whether people in household have or need adaptations. The questions were asked in 2014/15, 2019/20 and 2020/21}
\usage{
hh_adapt(folder, years)
}
\arguments{
\item{folder}{A folder containing EHS downloaded from UKDS and unzipped}

\item{years}{A list of years - the fewer and the more recent the less likely this is to return inconsistent data}
}
\description{
Import data on whether people in household have or need adaptations. The questions were asked in 2014/15, 2019/20 and 2020/21
}
