% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehs_sl.R
\name{hh_concealed}
\alias{hh_concealed}
\title{Import special licence EHS households data on concealed households - to be then matched against other imported data}
\usage{
hh_concealed(
  folder,
  years,
  min_age,
  income_filter = "Off",
  definition = "narrow"
)
}
\arguments{
\item{folder}{A folder containing EHS downloaded from UKDS and unzipped}

\item{years}{A list of years - the fewer and the more recent the less likely this is to return inconsistent data}
}
\description{
Import special licence EHS households data on concealed households - to be then matched against other imported data
}
