% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehs_sl.R
\name{hh_detailed}
\alias{hh_detailed}
\title{Import special licence EHS households data - including selected variables from detailed datasets}
\usage{
hh_detailed(folder, years)
}
\arguments{
\item{folder}{A folder containing EHS downloaded from UKDS and unzipped}

\item{years}{A list of years - the fewer and the more recent the less likely this is to return inconsistent data}
}
\description{
Import special licence EHS households data - including selected variables from detailed datasets
}
