% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehs_sl.R
\name{hh_disabled}
\alias{hh_disabled}
\title{Import data on whether household members have impairments and have difficulties using wheelchairs, and join onto special licence data}
\usage{
hh_disabled(folder, years)
}
\arguments{
\item{folder}{A folder containing EHS downloaded from UKDS and unzipped}

\item{years}{A list of years - the fewer and the more recent the less likely this is to return inconsistent data}
}
\description{
Import data on whether household members have impairments and have difficulties using wheelchairs, and join onto special licence data
}
